Introduce a number picker using the new GUI framework and use it in the options menu.
This change required a few large changes:
  + Refactor the input adapter system to have individual listeners for
start and end actions
  + Refactor Rectangle to use composition
  + Introduce a Graphics class that's composed of Slick's Graphics class

+++ java/believe/gui/GuiLayoutFactory.kt
@@ -20,7 +20,7 @@ class GuiLayoutFactory @Inject internal constructor(
      */
     inline fun <reified C, T : GuiElement> create(
         layoutBuilder: LayoutBuilder<C, T>, positionData: Rectangle = Rectangle(
-            x = 0, y = 0, width = guiContext.width, height = guiContext.height
+            x = 0f, y = 0f, width = guiContext.width.toFloat(), height = guiContext.height.toFloat()
         )
     ): T = if (Unit is C) {
         layoutBuilder.build(Unit, this, positionData)
